release 66.0 update LibVLC libraries

+++ resources/libvlc/include/vlc/libvlc_media.h
@@ -26,10 +26,6 @@
 #ifndef VLC_LIBVLC_MEDIA_H
 #define VLC_LIBVLC_MEDIA_H 1
 
-//#ifdef WIN32
-//#include <EABase/eabase.h>
-//#endif
-
 # ifdef __cplusplus
 extern "C" {
 # endif
@@ -369,7 +365,7 @@ typedef int (*libvlc_media_open_cb)(void *opaque, void **datap,
  * In particular, the callback should return an error if playback is stopped;
  * if it does not return, then libvlc_media_player_stop() will never return.
  */
-typedef int (*libvlc_media_read_cb)(void *opaque, unsigned char *buf,
+typedef ssize_t (*libvlc_media_read_cb)(void *opaque, unsigned char *buf,
                                         size_t len);
 
 /**
